requirements/tests.txt
-r base.txt

factory_boy==3.2.1
pytest==7.1.1
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-factoryboy==2.1.0
tox==3.24.5